requirements.txt
flask==2.3.3
flask-session==0.5.0
python-dotenv==1.0.0
openai>=1.0.0
pymupdf==1.23.5
requests==2.31.0
beautifulsoup4==4.12.2
werkzeug==2.3.7
httpx==0.27.2
python-docx>=0.8.11
python-pptx>=0.6.21
gunicorn==20.1.0
azure-ai-documentintelligence